producer/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
autopep8==2.0.4
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
confluent-kafka==2.2.0
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
fastapi==0.104.0
fonttools==4.43.1
greenlet==3.0.0
h11==0.14.0
httptools==0.6.1
idna==3.4
kiwisolver==1.4.5
lxml==4.9.3
matplotlib==3.8.0
numpy==1.26.1
obspy==1.4.0
packaging==23.2
Pillow==10.1.0
pycodestyle==2.11.1
pydantic==2.4.2
pydantic_core==2.10.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
scipy==1.11.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
uvloop==0.18.0
watchfiles==0.21.0
websockets==12.0
prometheus-client
psutil
